moving class. Nothing to see here
